requirements.txt
Flask==3.0.3
Werkzeug>=3.0.0
Pillow==10.4.0
gunicorn==22.0.0
